+++ src/neptune_fetcher/progress_update_handler.py
@@ -23,40 +23,70 @@ from abc import ABC
 
 
 class ProgressUpdateHandler(ABC):
-    def series_setup(self, total_series: int, series_limit: int) -> None:
+    """Abstract base class for progress update handlers.
+
+    Example usage:
+    >>> from neptune_fetcher import ReadOnlyProject
+    >>> from neptune_fetcher.progress_update_handler import ProgressUpdateHandler
+
+    >>> class CustomProgressUpdateHandler(ProgressUpdateHandler):
+    >>>     ...  # overwrite the methods you need
+
+    >>> project = ReadOnlyProject(...)
+    >>> project.progress_indicator(CustomProgressUpdateHandler())  # use your custom handler to track progress
+    """
+
+    def pre_series_fetch(self, total_series: int, series_limit: int) -> None:
+        """Executes before a series is fetched. Use it to set up the tracking up.
+
+        Args:
+            total_series: Total number of items in the series.
+            series_limit: Limit of items fetched in a single iteration.
+        """
         ...
 
-    def table_setup(self) -> None:
+    def pre_runs_table_fetch(self) -> None:
+        """Executes before a run table is fetched. Use it to set the tracking up."""
         ...
 
     def pre_download(self, total_size: int) -> None:
-        """Runs before a file or a fileset is downloaded. Use it to set the tracking up.
+        """Executes before a file or a fileset is downloaded. Use it to set the tracking up.
 
-        Parameters:
-            :param total_size: Total size of the file."""
+        Args:
+            total_size: Total size of the file.
+        """
         ...
 
     def on_series_fetch(self, step: int) -> None:
+        """Executes after every iteration of a series fetch. Use it to update the progress.
+
+        Args:
+            step: number of items that were fetched during the iteration."""
+        ...
+
+    def on_runs_table_fetch(self, step: int) -> None:
+        """Executes after every iteration of a runs table fetch. Use it to update the progress.
+
+        Args:
+            step: number of items that were fetched during the iteration."""
         ...
 
     def post_series_fetch(self) -> None:
+        """Executes after the series fetch is completed. Use it to clean the tracking up."""
         ...
 
-    def on_run_table_fetch(self, step: int) -> None:
-        ...
+    def post_runs_table_fetch(self) -> None:
+        """Executes after the run table fetch is completed. Use it to clean the tracking up."""
 
     def on_download_chunk(self, chunk: int) -> None:
-        """Runs after every iteration of a file or fileset download. Use it to update the progress.
+        """Executes after every iteration of a file or fileset download. Use it to update the progress.
 
-        Parameters:
-            :param chunk: number of bytes that were fetched during the iteration."""
-        ...
-
-    def post_table_fetch(self) -> None:
+        Args:
+            chunk: Number of bytes that were fetched during the iteration."""
         ...
 
     def post_download(self) -> None:
-        """Runs after the download is completed. Use it to clean the tracking up."""
+        """Executes after the download is completed. Use it to clean the tracking up."""
         ...
 
 
@@ -65,13 +95,13 @@ class NullProgressUpdateHandler(ProgressUpdateHandler):
 
 
 class DefaultProgressUpdateHandler(ProgressUpdateHandler):
-    def series_setup(self, total_series: int, series_limit: int) -> None:
+    def pre_series_fetch(self, total_series: int, series_limit: int) -> None:
         from tqdm import tqdm
 
         self._series_bar = tqdm(total=total_series)
         self._series_bar.update(n=series_limit)
 
-    def table_setup(self) -> None:
+    def pre_runs_table_fetch(self) -> None:
         from tqdm import tqdm
 
         self._table_bar = tqdm()
@@ -83,11 +113,11 @@ class DefaultProgressUpdateHandler(ProgressUpdateHandler):
     def post_series_fetch(self) -> None:
         self._series_bar.close()
 
-    def on_run_table_fetch(self, step: int) -> None:
+    def on_runs_table_fetch(self, step: int) -> None:
         self._table_bar.update(n=step)
         self._table_bar.set_description("Fetching runs")
 
-    def post_table_fetch(self) -> None:
+    def post_runs_table_fetch(self) -> None:
         self._table_bar.close()
 
     def pre_download(self, total_size: int) -> None: